change log and config hazelcast

+++ src/main/java/com/venesa/repository/LogRepository.java
@@ -1,12 +1,11 @@
 package com.venesa.repository;
 
+import com.venesa.entity.LogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.venesa.entity.LogEntity;
-
 @Repository
-public interface LogRepository extends JpaRepository<LogEntity, Integer>{
+public interface LogRepository extends JpaRepository<LogEntity, Long>{
 	
 
 }
